Add client code, add panic interface, add build tooling
Added all the client code from the panic-server side since things are moving out into different modules.

Now, when requiring panic-client, an interface object is exported with a "connect" method (and more to come) that syncs the client with the testing server.

Gulp is now in charge of compiling the src files into node/browser code.