Add client UID, invoke test runner
A good portion of the boilerplate client code has been finished (hopefully). After trying the persisten UID many different ways, I gave up and realized having a UID is more trouble than it's worth, and if it is worth it, it can be built as an extension.

Now when the "run" event happens, the runner function is invoked with the Test Description Object. That marks the full implementation.