Add travis file, add async discourse model, clean subscription logic
Now, instead of static test runners, the async model is implemented: you recieve a test, and other middleware can intercept, but eventually a `ready` event will be fired along with the test ID to transmit to.

Sadly, all tests are failing, but there is now a travis file to make it more obvious...

The incoming event system was dirty before the async model was fully implemented, so some things were more spaghetti like. Since then, the changes have allowed for a cleaner structure.