Add ability to fail tests, fail on catch, add comments. Tests failing.
Now the error property is added to termination events. If your code throws an error or manually fails the test, it will serialize and forward it to the server for parsing.

If a test throws an error while being parsed, it's caught and treated as a fail, and sent back to the server. This allows for other expectation suites to be used like `expect.js` and `should.js`, or just generic error catching.