Add eslint, remove index.html, bump version, move to mocha, remove obsolete code and replace it with a simpler model.
Eslint is now used in this project and uses the same settings as
panic-server.

The index.html file is not necessary and has been removed, along with a
bunch of other code.

Mocha is now used instead of jasmine.

A new model is now being used, one where there are two distinct actions:
1). Evaluate functions
2). Update state
This allows larger structures to be composed on top, instead of
wrangling everything into a test object, and also allows for behavior
like beforeEach.