Replace Job with new runner
The new runner has better async support, a slimmer API, and now reports
return (or promise-resolved) values. Big wins all around.
The Job function was largely replaced with `Context` & `runner`, but
there still needed to be an adapter layer between the runner and the
socket (testability and separation of concerns).